homework15 refactoring1.1 redesigning
