+++ bermanov-ivan/homework15/src/components/MessageField/MessageField.css
@@ -1,8 +1,3 @@
-.body {
-    height: 400px;
-    display: flex;
-}
-
 .message-field {
     width: 70%;
     display: flex;
